+++ tests/model/test_scvi.py
@@ -1122,7 +1122,7 @@ def test_scvi_no_anndata(n_batches: int = 3, n_latent: int = 5):
     datamodule.n_vars = adata.n_vars
     datamodule.n_batch = n_batches
 
-    model = SCVI(n_latent=5)
+    model = SCVI(n_latent=n_latent)
     assert model._module_init_on_train
     assert model.module is None
 
@@ -1140,7 +1140,7 @@ def test_scvi_no_anndata(n_batches: int = 3, n_latent: int = 5):
     datamodule.n_obs = 100_000_000  # large number for fewer default epochs
     model.train(datamodule=datamodule)
 
-    model = SCVI(adata, n_latent=5)
+    model = SCVI(adata, n_latent=n_latent)
     assert not model._module_init_on_train
     assert model.module is not None
     assert hasattr(model, "adata")
@@ -1170,7 +1170,7 @@ def test_scvi_no_anndata_with_external_indices(n_batches: int = 3, n_latent: int
     datamodule.n_vars = adata.n_vars
     datamodule.n_batch = n_batches
 
-    model = SCVI(n_latent=5)
+    model = SCVI(n_latent=n_latent)
     assert model._module_init_on_train
     assert model.module is None
 
@@ -1188,7 +1188,7 @@ def test_scvi_no_anndata_with_external_indices(n_batches: int = 3, n_latent: int
     datamodule.n_obs = 100_000_000  # large number for fewer default epochs
     model.train(datamodule=datamodule)
 
-    model = SCVI(adata, n_latent=5)
+    model = SCVI(adata, n_latent=n_latent)
     assert not model._module_init_on_train
     assert model.module is not None
     assert hasattr(model, "adata")